pyproject.toml
[project]
name = "social-mini"
version = "0.1.0"
description = "Социальная сеть мини-постов с аутентификацией и CRUD-функционалом"
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.13.3",
    "python-jose[cryptography]>=3.3.0",
    
    "passlib==1.7.4",
    "bcrypt==3.2.0",
    "python-multipart>=0.0.17",
    "email-validator>=2.1.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "httpx>=0.27.2"
]

[[project.authors]]
name = "Чепурных Артур"

[[project.authors]]
name = "Виктор Смирнов"

[[project.authors]]
name = "Илья Мельников"

[[project.authors]]
name = "Алена Рагозина"

[[project.authors]]
name = "Георгий Сердюков"

[[project.authors]]
name = "Анастасия Королева"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"